rename metric to sample/report
