+++ pkg/templates/build.sh
@@ -4,20 +4,20 @@ OUTPARAMS="Mgoogle/protobuf/duration.proto=github.com/golang/protobuf/ptypes/dur
 
 ## RUN PROTOC ON ALL THE TEMPLATES
 protoc mixer/pkg/templates/mixer/*.proto -I=. -I=api --go_out=${OUTPARAMS}
-protoc mixer/pkg/templates/metric/MetricTemplate.gen.proto mixer/pkg/templates/metric/MetricTemplate.proto -I=. -I=api --go_out=${OUTPARAMS}
+protoc mixer/pkg/templates/sample/report/ReportTesterTemplate.gen.proto mixer/pkg/templates/sample/report/ReportTesterTemplate.proto -I=. -I=api --go_out=${OUTPARAMS}
 
 ## SOME MAGIC TO MAKE COMPILER HAPPY
 sed -i \
   -e 's/ValueType_VALUE_TYPE_UNSPECIFIED/VALUE_TYPE_UNSPECIFIED/g' \
-  mixer/pkg/templates/metric/MetricTemplate.gen.pb.go;
+  mixer/pkg/templates/sample/report/ReportTesterTemplate.gen.pb.go;
 
 sed -i \
   -e 's/ValueType_VALUE_TYPE_UNSPECIFIED/VALUE_TYPE_UNSPECIFIED/g' \
-  mixer/pkg/templates/metric/MetricTemplate.pb.go;
+  mixer/pkg/templates/sample/report/ReportTesterTemplate.pb.go;
 
 ## BUILD INDIVIDUAL GENERATED PROCESSORS
 
-DIRS="metric"
+DIRS="sample/report"
 
 for pkgdir in ${DIRS}; do
     pushd mixer/pkg/templates/${pkgdir} > /dev/null; \
