+++ pkg/templates/sample/report/BUILD
