+++ pkg/templates/sample/report/ReportTesterTemplate.proto
@@ -1,12 +1,12 @@
 syntax = "proto3";
 
-package istio.mixer.adapter.metric;
+package istio.mixer.adapter.sample.report;
 
 import "mixer/v1/config/descriptor/value_type.proto";
 import "mixer/pkg/templates/mixer/TemplateExtensions.proto";
 
 option (istio.mixer.v1.config.template.template_variety) = TEMPLATE_VARIETY_REPORT;
-option (istio.mixer.v1.config.template.template_name) = "Metric";
+option (istio.mixer.v1.config.template.template_name) = "Sample";
 
 message Constructor {
     istio.mixer.v1.config.descriptor.ValueType value = 2;
